minor modular changes
