+++ app/src/main/java/com/bruhascended/fitapp/ui/addfood/AddFoodActivity.kt
@@ -8,7 +8,7 @@ import android.view.MenuItem
 import androidx.databinding.DataBindingUtil
 import com.bruhascended.fitapp.R
 import com.bruhascended.fitapp.databinding.ActivityAddFoodBinding
-import com.bruhascended.fitapp.ui.util.setupToolbar
+import com.bruhascended.fitapp.util.setupToolbar
 
 class AddFoodActivity : AppCompatActivity() {
