+++ app/src/main/java/com/bruhascended/fitapp/ui/addfood/FoodSearchActivity.kt
@@ -12,7 +12,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.bruhascended.fitapp.R
 import com.bruhascended.fitapp.databinding.ActivityFoodSearchBinding
-import com.bruhascended.fitapp.ui.util.setupToolbar
+import com.bruhascended.fitapp.util.setupToolbar
 import com.bruhascended.api.models.foods.Food
 import com.bruhascended.fitapp.ui.capturefood.PredictionPresenter
 import kotlinx.coroutines.*
